[future] Use weak_ptr<> in some callback self-captures, to avoid ref cycles.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/36502

I discovered via logging that sometimes we're deleting futures without
completing them.

This can be problematic from a memory leak perspective, since if the callback
lambda acquires a strong reference to the calling future, it will leak memory -
not just the lambda involved, but also the other strong references.

This change simply switches to capturing a weak pointer instead of a strong
one for the relevant addCallback() futures.
ghstack-source-id: 102058445

Differential Revision: [D20998905](https://our.internmc.facebook.com/intern/diff/D20998905/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D20998905/)!